Rename packages `columns` -> `column`, `fields` -> `field`
