+++ tools/protoc-plugin/src/main/java/io/spine/tools/protoc/Plugin.java
@@ -25,8 +25,8 @@ import com.google.protobuf.compiler.PluginProtos.CodeGeneratorRequest;
 import com.google.protobuf.compiler.PluginProtos.CodeGeneratorResponse;
 import io.spine.code.proto.OptionExtensionRegistry;
 import io.spine.tools.protoc.builder.BuilderGenerator;
-import io.spine.tools.protoc.columns.ColumnGenerator;
-import io.spine.tools.protoc.fields.FieldGenerator;
+import io.spine.tools.protoc.column.ColumnGenerator;
+import io.spine.tools.protoc.field.FieldGenerator;
 import io.spine.tools.protoc.iface.InterfaceGenerator;
 import io.spine.tools.protoc.method.MethodGenerator;
 import io.spine.tools.protoc.nested.NestedClassGenerator;
