+++ tools/protoc-plugin/src/main/java/io/spine/tools/protoc/field/FieldGenerationTask.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.tools.protoc.fields;
+package io.spine.tools.protoc.field;
 
 import com.google.common.collect.ImmutableList;
 import io.spine.code.gen.java.FieldFactory;
