+++ tools/protoc-plugin/src/main/java/io/spine/tools/protoc/field/GenerateFieldsByPattern.java
@@ -1,4 +1,4 @@
-package io.spine.tools.protoc.fields;
+package io.spine.tools.protoc.field;
 
 import com.google.common.collect.ImmutableList;
 import io.spine.code.gen.java.FieldFactory;
