+++ tools/protoc-plugin/src/main/java/io/spine/tools/protoc/field/package-info.java
@@ -22,7 +22,7 @@
 
 @CheckReturnValue
 @ParametersAreNonnullByDefault
-package io.spine.tools.protoc.fields;
+package io.spine.tools.protoc.field;
 
 import com.google.errorprone.annotations.CheckReturnValue;
